Move and implement atomic design
