+++ purchase-app/src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Footer from "../components/footer";
-import Navbar from "../components/navbar";
+import Navbar from "../components/organisms/Navbar";
 
 const inter = Inter({ subsets: ["latin"] });
