+++ purchase-app/src/app/page.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Container  from "@/components/container";
+import Container  from "@/components/molecules/Container";
 
 export default function Home() {
   return (
