+++ purchase-app/src/components/organisms/Footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Container from '@/components/container';
+import Container from '@/components/molecules/Container';
 
 const Footer = () => {
   return (
