+++ purchase-app/src/components/organisms/Navbar.tsx
@@ -1,16 +1,22 @@
-import React from 'react';
+"use client"
 import Link from 'next/link';
+import TextInput from '../molecules/TextInput';
 
 const Navbar = () => {
+  const handleChangeInput = (value: string) => {
+    console.log("🚀 ~ handleChangeInput ~ value:", value)
+  }
+
   return (
-    <nav className="navbar fixed top-0 left-0 w-full bg-white shadow-md z-10">
-      <div className="container mx-auto py-4 px-6 flex justify-between items-center">
+    <nav id='navbar' className="w-full bg-white text-black shadow-md">
+      <div id='navbar-container' className="mx-auto py-4 px-6 flex justify-between items-center">
         <Link href="/">
             <button className='text-3xl font-bold focus:outline-none'>
                 Ecoomerce
             </button>
         </Link>
-        <form action="/pages/search" method="GET" className="flex-grow flex items-center justify-center">
+        <TextInput onChange={handleChangeInput} />
+        {/* <form action="/pages/search" method="GET" className="flex-grow flex items-center justify-center">
           <input
             type="text"
             placeholder="Search products..."
@@ -20,7 +26,7 @@ const Navbar = () => {
           <button type="submit" className="bg-blue-500 text-white px-4 py-1 rounded-full ml-2">
             Search
           </button>
-        </form>
+        </form> */}
         <div className="flex items-center space-x-4">
           <Link href="/pages/account/login" passHref>
             <button className="btn btn-success">Account</button>
